More standardization on IF package name.
