+++ safe/impact_functions/inundation/flood_polygon_population/test/test_flood_polygon_population.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 """
 InaSAFE Disaster risk assessment tool developed by AusAid and World Bank
-- **Test for Flood Population Evacuation Raster Impact Function.**
+- *Flood Vector on Population Test Cases.**
 
 Contact : ole.moller.nielsen@gmail.com
 
@@ -22,27 +22,25 @@ import unittest
 from safe.storage.core import read_layer
 from safe.impact_functions.impact_function_manager \
     import ImpactFunctionManager
-from safe.impact_functions.inundation\
-    .flood_population_evacuation_raster_hazard.impact_function import \
-    FloodEvacuationRasterHazardFunction
 from safe.test.utilities import get_qgis_app, test_data_path
-from safe.common.utilities import OrderedDict
+from safe.impact_functions.inundation.flood_polygon_population\
+    .impact_function import FloodEvacuationVectorHazardFunction
 
 QGIS_APP, CANVAS, IFACE, PARENT = get_qgis_app()
 
 
-class TestFloodEvacuationFunctionRasterHazard(unittest.TestCase):
+class TestFloodEvacuationVectorHazardFunction(unittest.TestCase):
     """Test for Flood Vector Building Impact Function."""
 
     def setUp(self):
         registry = ImpactFunctionManager().registry
         registry.clear()
-        registry.register(FloodEvacuationRasterHazardFunction)
+        registry.register(FloodEvacuationVectorHazardFunction)
 
     def test_run(self):
-        function = FloodEvacuationRasterHazardFunction.instance()
+        function = FloodEvacuationVectorHazardFunction.instance()
 
-        hazard_path = test_data_path('hazard', 'continuous_flood_20_20.asc')
+        hazard_path = test_data_path('hazard', 'flood_multipart_polygons.shp')
         exposure_path = test_data_path(
             'exposure', 'pop_binary_raster_20_20.asc')
         hazard_layer = read_layer(hazard_path)
@@ -53,35 +51,21 @@ class TestFloodEvacuationFunctionRasterHazard(unittest.TestCase):
         function.run()
         impact = function.impact
 
-        # Count of flooded objects is calculated "by the hands"
-        # print "keywords", keywords
         keywords = impact.get_keywords()
-        evacuated = float(keywords['evacuated'])
-        total_needs_full = keywords['total_needs']
-        total_needs_weekly = OrderedDict([
-            [x['table name'], x['amount']] for x in
-            total_needs_full['weekly']
-        ])
-        total_needs_single = OrderedDict([
-            [x['table name'], x['amount']] for x in
-            total_needs_full['single']
-        ])
+        # print "keywords", keywords
+        affected_population = float(keywords['affected_population'])
+        total_population = keywords['total_population']
 
-        expected_evacuated = 100
-        self.assertEqual(evacuated, expected_evacuated)
-        self.assertEqual(total_needs_weekly['Rice [kg]'], 280)
-        self.assertEqual(total_needs_weekly['Family Kits'], 20)
-        self.assertEqual(total_needs_weekly['Drinking Water [l]'], 1750)
-        self.assertEqual(total_needs_weekly['Clean Water [l]'], 6700)
-        self.assertEqual(total_needs_single['Toilets'], 5)
+        self.assertEqual(affected_population, 20)
+        self.assertEqual(total_population, 200)
 
     def test_filter(self):
         """Test filtering IF from layer keywords"""
         hazard_keywords = {
             'subcategory': 'flood',
-            'unit': 'metres_depth',
-            'layer_type': 'raster',
-            'data_type': 'continuous'
+            'unit': 'wetdry',
+            'layer_type': 'vector',
+            'data_type': 'polygon'
         }
 
         exposure_keywords = {
@@ -98,7 +82,7 @@ class TestFloodEvacuationFunctionRasterHazard(unittest.TestCase):
 
         retrieved_if = impact_functions[0].metadata().as_dict()['id']
         expected = ImpactFunctionManager().get_function_id(
-            FloodEvacuationRasterHazardFunction)
+            FloodEvacuationVectorHazardFunction)
         message = 'Expecting %s, but getting %s instead' % (
             expected, retrieved_if)
         self.assertEqual(expected, retrieved_if, message)
