+++ safe/impact_functions/inundation/flood_raster_population/test/test_flood_raster_population.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 """
 InaSAFE Disaster risk assessment tool developed by AusAid and World Bank
-- *Flood Vector on Population Test Cases.**
+- **Test for Flood Population Evacuation Raster Impact Function.**
 
 Contact : ole.moller.nielsen@gmail.com
 
@@ -22,26 +22,26 @@ import unittest
 from safe.storage.core import read_layer
 from safe.impact_functions.impact_function_manager \
     import ImpactFunctionManager
+from safe.impact_functions.inundation.flood_raster_population.impact_function\
+    import FloodEvacuationRasterHazardFunction
 from safe.test.utilities import get_qgis_app, test_data_path
-from safe.impact_functions.inundation\
-    .flood_population_evacuation_polygon_hazard.impact_function import \
-    FloodEvacuationVectorHazardFunction
+from safe.common.utilities import OrderedDict
 
 QGIS_APP, CANVAS, IFACE, PARENT = get_qgis_app()
 
 
-class TestFloodEvacuationVectorHazardFunction(unittest.TestCase):
+class TestFloodEvacuationFunctionRasterHazard(unittest.TestCase):
     """Test for Flood Vector Building Impact Function."""
 
     def setUp(self):
         registry = ImpactFunctionManager().registry
         registry.clear()
-        registry.register(FloodEvacuationVectorHazardFunction)
+        registry.register(FloodEvacuationRasterHazardFunction)
 
     def test_run(self):
-        function = FloodEvacuationVectorHazardFunction.instance()
+        function = FloodEvacuationRasterHazardFunction.instance()
 
-        hazard_path = test_data_path('hazard', 'flood_multipart_polygons.shp')
+        hazard_path = test_data_path('hazard', 'continuous_flood_20_20.asc')
         exposure_path = test_data_path(
             'exposure', 'pop_binary_raster_20_20.asc')
         hazard_layer = read_layer(hazard_path)
@@ -52,21 +52,35 @@ class TestFloodEvacuationVectorHazardFunction(unittest.TestCase):
         function.run()
         impact = function.impact
 
-        keywords = impact.get_keywords()
+        # Count of flooded objects is calculated "by the hands"
         # print "keywords", keywords
-        affected_population = float(keywords['affected_population'])
-        total_population = keywords['total_population']
+        keywords = impact.get_keywords()
+        evacuated = float(keywords['evacuated'])
+        total_needs_full = keywords['total_needs']
+        total_needs_weekly = OrderedDict([
+            [x['table name'], x['amount']] for x in
+            total_needs_full['weekly']
+        ])
+        total_needs_single = OrderedDict([
+            [x['table name'], x['amount']] for x in
+            total_needs_full['single']
+        ])
 
-        self.assertEqual(affected_population, 20)
-        self.assertEqual(total_population, 200)
+        expected_evacuated = 100
+        self.assertEqual(evacuated, expected_evacuated)
+        self.assertEqual(total_needs_weekly['Rice [kg]'], 280)
+        self.assertEqual(total_needs_weekly['Family Kits'], 20)
+        self.assertEqual(total_needs_weekly['Drinking Water [l]'], 1750)
+        self.assertEqual(total_needs_weekly['Clean Water [l]'], 6700)
+        self.assertEqual(total_needs_single['Toilets'], 5)
 
     def test_filter(self):
         """Test filtering IF from layer keywords"""
         hazard_keywords = {
             'subcategory': 'flood',
-            'unit': 'wetdry',
-            'layer_type': 'vector',
-            'data_type': 'polygon'
+            'unit': 'metres_depth',
+            'layer_type': 'raster',
+            'data_type': 'continuous'
         }
 
         exposure_keywords = {
@@ -83,7 +97,7 @@ class TestFloodEvacuationVectorHazardFunction(unittest.TestCase):
 
         retrieved_if = impact_functions[0].metadata().as_dict()['id']
         expected = ImpactFunctionManager().get_function_id(
-            FloodEvacuationVectorHazardFunction)
+            FloodEvacuationRasterHazardFunction)
         message = 'Expecting %s, but getting %s instead' % (
             expected, retrieved_if)
         self.assertEqual(expected, retrieved_if, message)
